BreakerBall Tutorial
This is a simple project for use with a 96Boards 96Boards DragonBoard
and a Sensors Mezzanine card. The project shows how to build and upload
to the sensors mezzanine card, then how to recieve information from a
sensor and how to pass the information to a higher level program written
in Scratch.